e2b_client_example/pyproject.toml
[project]
name = "e2b-client-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "e2b>=2.10.2",
    "e2b-code-interpreter>=2.4.1",
    "python-dotenv>=1.2.1",
]
